RavenDB-23361 Moved test to separate file and added assertion
